rename jwt module to eth_sign
